mvin: a work in progress.  adds another commandline option; shellrc.  I started going down this road, but now I'm having second thoughts--namely, that this really isn't the purpose of this whole project.  The purpose is to easily make dotfiles somewhat portable between machines; not patching and updating the users shell environment files.  People can do what they wish.  Not sure this is worth it.
I am about to take a DRASTICLY different approach with this program.  Changing it command line arguments, behaviour and directory structure.  I don't really want to do much on this current branch (really, currently is trunk now, but soon to be a branch/tag---to never be (at this point at least), returned to again.).

Therefore, just wanted to say I haven't worked on this in a couple weeks and thus cannot guarantee the stability, correctness, yada yada of this program.  Use at own risk.


git-svn-id: svn://svn/de/trunk/proj/mvin@1023 a64eb3ba-58fa-4d19-8208-1a35c998e62f