bash_user env entries split out to files... aliases, functions, and variables
MORE NFO
Coming to you from sunny 2009-10-16!

There is a pretty good amount of reorg going on in this commit.  I tryed
to recall back to 2009 and wrote about it as best I could, below.

The .bash_user file--
Oh, the .bash_user file.  Well... this file was essentially mandated for
use by The Powers That Be--it HAD to be copied from
{release sw}/bash_user_dev.env to $HOME/.bash_user in order for the
devel environment to be init'd correctly.

Also, please note that the .bash_user file is near the top of user's
environment initialization call stack (by the bash shell) i.e. it's like
the first thing that is called whenever you log in to your devel
machine.  I want to handle/or have more control over how my personal
environment is init'd.  So I just wrote my own little stupid top-level
.bash_user file, then sourced bash_user_dev.env at the top of it.

Description of .dotfiles/bash_user_dev.env* files--
* bash_user_dev.env: local working copy
* bash_user_dev.env.patch : generated during patch application (rubbish)
* bash_user_dev.env.patch.README : cmds to patch
* bash_user_dev.env.production : symlink to latest release

Reasoning for creating the .dotfiles/bash_user_dev.env* files--
Every time The Powers That Be released an updated bash_user_dev.env
file, the onus was on you to get it.  If you didn't, well that's just
another "possibility" (or variable (or another thing to possibly go
wrong)) that you have to *remember* to contend with if you start getting
weird issues, somethings weren't working, general debugging of some
changes you made, etc.--when, to begin with, all the problem ever was, was
that you were not running the most up2date version of bash_user_dev.env.
Yahtzee.