Revert "cleanup: already have this dotfile (confirmed)"
This reverts commit 4e99fa25658a69970e2e0043f94d5753c3834014.

This reverts the commit that deletes the .screenrc file that was used on
walrus.

The reason it was deleted before was because its a duplicate / already
have this screenrc's contents accounted for elsewhere.

*HOWEVER*

The reason I am restoring this is because I would like to have 2 walrus
folders side-by-side.  One will have a lot of junk and stuff and the
other will be the cleaned up and, most likely, the
'never-to-be-touched-again' version.