walrus.essential: creates 'walrus.essential'-to contain essentials of 'walrus'.
MORE NFO
adds 'walrus.essential' (copy of 'walrus') right next to the
original 'walrus' folder.

Do this because I would like to have both side-by-side, with one
containing a lot of junk, 'walrus', and the other will be the
cleaned up, minimized version.

Most likely, both will be versions that are
'never-to-be-touched-again' version.