ec2: redefines terminal prompt according to ~/.whereami content
This is particularly useful on ec2 vm's where the hostname portion of
the terminal prompt is basically useless, at least for me.  Useful if
you need to do internal ec2 trasfers or something... you have the LAN ip
(or "hostname" rather (ugly one at that)) right there for you.

But if you're not doing ec2-to-ec2 transfers 24/7 via their LAN, ITS
COMPLETELY USELESS, so this change attempts to make it better. This
approach works great for me.  Whenever log in to ec2 instance the first
time, just scp these 2 files to $HOME, and update ~/.whereami to
whatever you want your terminal prompt to be (usually the DNS hostname
for me at least).