super important binary for controlling bump's cooling system (bump)
when a past updated release was applied, the ipmitool that shipped with
was *unable* to control the system's cooling.  thankfully, had a backup.
a few weeks ago I thought I had lost this guy but huehue was just that
the pool was not mounted; still had it.

putting this binary in dotfiles just as another backup place since its
small and stuff.